refactor: move package debug errors
